reorganize single-file into packages
